[#1440] About redesign
