+++ ui-design-lib/src/main/res/ui/common/drawable/zashi_text_logo_small.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="64dp"
+    android:height="18dp"
+    android:viewportWidth="64"
+    android:viewportHeight="18">
+  <path
+      android:pathData="M19.74,1.408L23.015,0.916L28.778,17.572L25.031,17.913M25.031,17.885L23.376,13.622L17.753,13.292L16.098,17.583L13.453,17.887L19.737,1.41M18.746,10.984L22.352,10.989L20.398,6.031L18.744,10.982L18.746,10.984Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M35.2,17.887C32.684,17.887 30.372,17.152 29.045,15.381L32.234,12.691C33.178,13.902 34.044,14.691 35.504,15.273L38.412,15.244C38.681,14.729 38.794,14.236 38.794,13.765C38.794,12.241 37.985,11.77 34.976,11.186C31.54,10.536 29.427,9.305 29.427,6.523C29.427,5.515 29.944,4.304 30.751,3.431L36.052,0.913C38.275,0.913 40.769,1.681 42.093,3.452L38.903,6.141C37.959,4.931 37.017,4.17 35.556,3.586H32.816C32.569,4.012 32.434,4.483 32.434,4.954C32.434,6.389 33.444,6.836 36.701,7.42C40.385,8.026 41.822,9.37 41.822,12.039C41.822,13.047 41.374,14.234 40.633,15.334L35.198,17.885L35.2,17.887Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M53.374,0.916L57.263,1.408V17.583L53.374,17.885V10.426L47.339,10.32V17.581L43.37,17.885V1.37L47.339,0.916V8.01L53.374,7.96V0.916Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M59.25,1.363L63.22,0.917V17.887L59.25,17.585V1.363Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M5.74,13.85H13.349L12.013,17.496H0.78V16.49L8.719,4.606H1.441L2.432,1.304H13.677"
+      android:fillColor="#ffffff"/>
+</vector>
