+++ ui-lib/src/main/res/ui/about/values/strings.xml
@@ -1,19 +1,17 @@
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="about_title">About</string>
-    <string name="about_version_format" formatted="true">Version <xliff:g example="1.0" id="version_name">%1$s
-    </xliff:g></string>
+    <string name="about_version_format" formatted="true">Zashi Version %s</string>
     <string name="about_debug_menu_app_name">App name:
         <xliff:g example="Zashi" id="app_name">%1$s</xliff:g></string>
     <string name="about_debug_menu_build">Build: <xliff:g example="635dac0eb9ddc2bc6da5177f0dd495d8b76af4dc"
         id="git_commit_hash">%1$s</xliff:g></string>
-    <string name="about_description">Send and receive ZEC on Zashi! Zashi is a minimal-design, self-custody, ZEC-only
-        shielded wallet that keeps your transaction history and wallet balance private. Built by Zcashers, for
-        Zcashers. Developed and maintained by Electric Coin Co., the inventor of Zcash, Zashi features a built-in
-        user-feedback mechanism to enable more features, more quickly.</string>
+    <string name="about_description">Send and receive ZEC on Zashi!\nZashi is a minimal-design, self-custody, ZEC-only
+        shielded wallet that keeps your transaction history and wallet balance private. Built by Zcashers, for Zcashers. Developed and maintained by Electric Coin Co., the inventor of Zcash, Zashi features a built-in user-feedback mechanism to enable more features, more quickly.</string>
 
     <string name="about_pp_part_1">See our Privacy Policy\u0020</string>
     <string name="about_pp_part_2">here</string>
     <string name="about_pp_part_3">.</string>
     <string name="about_unable_to_web_browser">Unable to find a web browser app.</string>
-    <string name="about_button_whats_new">What\'s new</string>
+    <string name="about_button_whats_new">WHAT\'S NEW</string>
+    <string name="about_button_privacy_policy">PRIVACY POLICY</string>
 </resources>
